- refactor types into separate definition files - fix login fail handling type error
