+++ src/lib/types/artifact-types.d.ts
