+++ src/lib/types/workspace-types.d.ts
@@ -1,0 +1,32 @@
+
+export interface AdditionalSpecification {
+  id: string;
+  content: string;
+}
+
+export interface Specification {
+  id: string;
+  name: string;
+  topic: string;
+  count?: number; // for quiz only
+  writingLevel: string;
+  comprehensionLevel: string;
+  additionalSpecs: AdditionalSpecification[];
+}
+
+export interface Page {
+  id: string;
+  title: string;
+  content: string;
+}
+
+export interface Workspace {
+  id: string;
+  name: string;
+  fileUrls: string[];
+  createdAt: number;
+  locked?: boolean;
+  materialType: string;
+  specifications: Specification[];
+  pages: Page[];
+}
