+++ src/redux/slices/workspaceSlice.ts
@@ -1,42 +1,11 @@
 // redux/workspaceSlice.ts
 import { createSlice, PayloadAction, createAsyncThunk } from '@reduxjs/toolkit';
 import { RootState } from '../store';
-import { GET as getMaterials } from '@/app/api/material/route';
-import { GET as _getSpecifications } from "@/app/api/material/specification/route";
-import { POST as _addLessonPage, GET as _getLessonPages } from "@/app/api/material/page/route";
+import { GET as getMaterials } from '@/app/api/workspace/route';
+import { GET as _getSpecifications } from "@/app/api/workspace/specification/route";
+import { POST as _addLessonPage, GET as _getLessonPages } from "@/app/api/workspace/page/route";
 import RequestBuilder from '@/lib/hooks/builders/request-builder';
-
-export interface AdditionalSpecification {
-  id: string;
-  content: string;
-}
-
-export interface Specification {
-  id: string;
-  name: string;
-  topic: string;
-  count?: number; // for quiz only
-  writingLevel: string;
-  comprehensionLevel: string;
-  additionalSpecs: AdditionalSpecification[];
-}
-
-export interface Page {
-  id: string;
-  title: string;
-  content: string;
-}
-
-export interface Workspace {
-  id: string;
-  name: string;
-  fileUrls: string[];
-  createdAt: number;
-  locked?: boolean;
-  materialType: string;
-  specifications: Specification[];
-  pages: Page[];
-}
+import { Page, Specification, Workspace } from '@/lib/types/workspace-types';
 
 interface WorkspaceState {
   workspaces: Workspace[];
